ios: Fix activity backfill page limit type
Use an Int page-limit constant for getSessionMessagesPageBuffer so
iOS release archives and TestFlight uploads build successfully.

+++ ios/DelightApp/SDKBridge.swift
@@ -391,7 +391,7 @@ private enum UpdateTiming {
     static let sessionActivityBackfillDelaySeconds: TimeInterval = 0.2
     static let sessionActivityBackfillMinIntervalSeconds: TimeInterval = 5.0
     static let sessionActivityBackfillMaxSessionsPerPass: Int = 20
-    static let sessionActivityBackfillPageLimit: Int64 = 20
+    static let sessionActivityBackfillPageLimit: Int = 20
     static let pairingDiscoveryPollIntervalSeconds: TimeInterval = 2.0
     static let pairingDiscoveryMaxAttempts: Int = 30
 }
